Desktop/NRF_SEARCH_POC/fashion-agent-python-backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]==0.24.0
pydantic>=2.9.0
anthropic>=0.18.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
google-genai>=1.0.0
pillow>=10.0.0
starlette>=0.41.3
